@@ -1,9 +1,9 @@
 package com.tom.storagemod;
 
-import net.minecraft.block.Block;
 import net.minecraft.tags.BlockTags;
-import net.minecraft.tags.ITag;
+import net.minecraft.tags.Tag.Named;
+import net.minecraft.world.level.block.Block;
 
 public class StorageTags {
-	public static final ITag.INamedTag<Block> REMOTE_ACTIVATE = BlockTags.bind("toms_storage:remote_activate");
+	public static final Named<Block> REMOTE_ACTIVATE = BlockTags.bind("toms_storage:remote_activate");
 }
