@@ -2,23 +2,23 @@ package com.tom.storagemod.item;
 
 import java.util.List;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.ItemUseContext;
-import net.minecraft.item.Items;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.nbt.NBTUtil;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.util.text.IFormattableTextComponent;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.world.World;
+import net.minecraft.ChatFormatting;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.NbtUtils;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
 
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
@@ -26,8 +26,6 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import com.tom.storagemod.StorageMod;
 import com.tom.storagemod.block.IPaintable;
 import com.tom.storagemod.proxy.ClientProxy;
-
-import net.minecraft.item.Item.Properties;
 
 public class ItemPaintKit extends Item {
 
@@ -38,31 +36,31 @@ public class ItemPaintKit extends Item {
 
 	@Override
 	@OnlyIn(Dist.CLIENT)
-	public void appendHoverText(ItemStack stack, World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+	public void appendHoverText(ItemStack p_41421_, Level p_41422_, List<Component> tooltip, TooltipFlag p_41424_) {
 		ClientProxy.tooltip("paint_kit", tooltip);
 	}
 
 	@Override
-	public ActionResultType useOn(ItemUseContext context) {
+	public InteractionResult useOn(UseOnContext context) {
 		if(!context.getLevel().isClientSide) {
 			if(context.isSecondaryUseActive()) {
 				BlockState state = context.getLevel().getBlockState(context.getClickedPos());
-				TileEntity tile = context.getLevel().getBlockEntity(context.getClickedPos());
+				BlockEntity tile = context.getLevel().getBlockEntity(context.getClickedPos());
 				if(tile == null && state.canOcclude()) {
 					ItemStack is = context.getItemInHand();
-					if(!is.hasTag())is.setTag(new CompoundNBT());
-					is.getTag().put("block", NBTUtil.writeBlockState(state));
+					if(!is.hasTag())is.setTag(new CompoundTag());
+					is.getTag().put("block", NbtUtils.writeBlockState(state));
 					//ITextComponent tc = new TranslationTextComponent("tooltip.toms_storage.set_paint", state.getBlock().getNameTextComponent().applyTextStyle(TextFormatting.GREEN));
 					//context.getPlayer().sendStatusMessage(tc, true);
 				}
-				return ActionResultType.SUCCESS;
+				return InteractionResult.SUCCESS;
 			} else {
 				BlockState state = context.getLevel().getBlockState(context.getClickedPos());
 				ItemStack is = context.getItemInHand();
 				if(is.hasTag() && is.getTag().contains("block") && state.getBlock() instanceof IPaintable) {
-					if(((IPaintable)state.getBlock()).paint(context.getLevel(), context.getClickedPos(), NBTUtil.readBlockState(is.getTag().getCompound("block")))) {
-						PlayerEntity playerentity = context.getPlayer();
-						context.getLevel().playSound(playerentity, context.getClickedPos(), SoundEvents.BUCKET_EMPTY, SoundCategory.BLOCKS, 1.0F, 1.0F);
+					if(((IPaintable)state.getBlock()).paint(context.getLevel(), context.getClickedPos(), NbtUtils.readBlockState(is.getTag().getCompound("block")))) {
+						Player playerentity = context.getPlayer();
+						context.getLevel().playSound(playerentity, context.getClickedPos(), SoundEvents.BUCKET_EMPTY, SoundSource.BLOCKS, 1.0F, 1.0F);
 						if(playerentity != null) {
 							is.hurtAndBreak(1, context.getPlayer(), p -> p.broadcastBreakEvent(context.getHand()));
 							if(is.isEmpty()) {
@@ -70,26 +68,26 @@ public class ItemPaintKit extends Item {
 							}
 						}
 					}
-					return ActionResultType.SUCCESS;
+					return InteractionResult.SUCCESS;
 				}
 			}
 		} else {
 			BlockState state = context.getLevel().getBlockState(context.getClickedPos());
-			if(context.isSecondaryUseActive())return ActionResultType.SUCCESS;
+			if(context.isSecondaryUseActive())return InteractionResult.SUCCESS;
 			if(state.getBlock() instanceof IPaintable) {
-				return ActionResultType.SUCCESS;
+				return InteractionResult.SUCCESS;
 			}
 		}
 		return super.useOn(context);
 	}
 
 	@Override
-	public ITextComponent getName(ItemStack is) {
-		IFormattableTextComponent tc = (IFormattableTextComponent) super.getName(is);
+	public Component getName(ItemStack is) {
+		MutableComponent tc = (MutableComponent) super.getName(is);
 		if(is.hasTag() && is.getTag().contains("block")) {
-			BlockState st = NBTUtil.readBlockState(is.getTag().getCompound("block"));
+			BlockState st = NbtUtils.readBlockState(is.getTag().getCompound("block"));
 			tc.append(" (");
-			tc.append(st.getBlock().getName().withStyle(TextFormatting.GREEN));
+			tc.append(st.getBlock().getName().withStyle(ChatFormatting.GREEN));
 			tc.append(")");
 		}
 		return tc;
