@@ -1,8 +1,8 @@
 package com.tom.storagemod.jei;
 
-import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.nbt.CompoundTag;
 
 public interface IJEIAutoFillTerminal {
-	void sendMessage(CompoundNBT compound);
+	void sendMessage(CompoundTag compound);
 
 }
