@@ -1,7 +1,7 @@
 package com.tom.storagemod.network;
 
-import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.nbt.CompoundTag;
 
 public interface IDataReceiver {
-	void receive(CompoundNBT tag);
+	void receive(CompoundTag tag);
 }
