@@ -1,8 +1,8 @@
 package com.tom.storagemod.tile;
 
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 
 import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.items.IItemHandler;
@@ -11,8 +11,8 @@ public class InfoHandler implements IItemHandler {
 	private static final ItemStack STACK = new ItemStack(Items.BARRIER, Integer.MAX_VALUE);
 	public static final LazyOptional<IItemHandler> INSTANCE = LazyOptional.of(InfoHandler::new);
 	static {
-		STACK.setTag(new CompoundNBT());
-		CompoundNBT d = new CompoundNBT();
+		STACK.setTag(new CompoundTag());
+		CompoundTag d = new CompoundTag();
 		STACK.getTag().put("display", d);
 		d.putString("Name", "{\"translate\":\"tooltip.toms_storage.loop\",\"color\":\"red\",\"italic\":false}");
 	}
