@@ -4,20 +4,20 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.IntStream;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.inventory.InventoryHelper;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.inventory.container.INamedContainerProvider;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.tileentity.ITickableTileEntity;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.tileentity.TileEntityType;
-import net.minecraft.util.Direction;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.Containers;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
 
 import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.items.CapabilityItemHandler;
@@ -27,33 +27,34 @@ import net.minecraftforge.items.ItemHandlerHelper;
 import com.tom.storagemod.Config;
 import com.tom.storagemod.StorageMod;
 import com.tom.storagemod.StoredItemStack;
+import com.tom.storagemod.TickerUtil.TickableServer;
 import com.tom.storagemod.block.StorageTerminalBase;
 import com.tom.storagemod.block.StorageTerminalBase.TerminalPos;
 import com.tom.storagemod.gui.ContainerStorageTerminal;
 import com.tom.storagemod.item.ItemWirelessTerminal;
 
-public class TileEntityStorageTerminal extends TileEntity implements INamedContainerProvider, ITickableTileEntity {
+public class TileEntityStorageTerminal extends BlockEntity implements MenuProvider, TickableServer {
 	private IItemHandler itemHandler;
 	private Map<StoredItemStack, Long> items = new HashMap<>();
 	private int sort;
 	private String lastSearch = "";
 	private boolean updateItems;
-	public TileEntityStorageTerminal() {
-		super(StorageMod.terminalTile);
+	public TileEntityStorageTerminal(BlockPos pos, BlockState state) {
+		super(StorageMod.terminalTile, pos, state);
 	}
 
-	public TileEntityStorageTerminal(TileEntityType<?> tileEntityTypeIn) {
-		super(tileEntityTypeIn);
+	public TileEntityStorageTerminal(BlockEntityType<?> tileEntityTypeIn, BlockPos pos, BlockState state) {
+		super(tileEntityTypeIn, pos, state);
 	}
 
 	@Override
-	public Container createMenu(int id, PlayerInventory plInv, PlayerEntity arg2) {
+	public AbstractContainerMenu createMenu(int id, Inventory plInv, Player arg2) {
 		return new ContainerStorageTerminal(id, plInv, this);
 	}
 
 	@Override
-	public ITextComponent getDisplayName() {
-		return new TranslationTextComponent("ts.storage_terminal");
+	public Component getDisplayName() {
+		return new TranslatableComponent("ts.storage_terminal");
 	}
 
 	public Map<StoredItemStack, Long> getStacks() {
@@ -102,19 +103,19 @@ public class TileEntityStorageTerminal extends TileEntity implements INamedConta
 		if(st.isEmpty())return;
 		StoredItemStack st0 = pushStack(new StoredItemStack(st));
 		if(st0 != null) {
-			InventoryHelper.dropItemStack(level, worldPosition.getX() + .5f, worldPosition.getY() + .5f, worldPosition.getZ() + .5f, st0.getActualStack());
+			Containers.dropItemStack(level, worldPosition.getX() + .5f, worldPosition.getY() + .5f, worldPosition.getZ() + .5f, st0.getActualStack());
 		}
 	}
 
 	@Override
-	public void tick() {
-		if(!level.isClientSide && updateItems) {
+	public void updateServer() {
+		if(updateItems) {
 			BlockState st = level.getBlockState(worldPosition);
 			Direction d = st.getValue(StorageTerminalBase.FACING);
 			TerminalPos p = st.getValue(StorageTerminalBase.TERMINAL_POS);
 			if(p == TerminalPos.UP)d = Direction.UP;
 			if(p == TerminalPos.DOWN)d = Direction.DOWN;
-			TileEntity invTile = level.getBlockEntity(worldPosition.relative(d));
+			BlockEntity invTile = level.getBlockEntity(worldPosition.relative(d));
 			items.clear();
 			if(invTile != null) {
 				LazyOptional<IItemHandler> lih = invTile.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, d.getOpposite());
@@ -128,7 +129,7 @@ public class TileEntityStorageTerminal extends TileEntity implements INamedConta
 		}
 	}
 
-	public boolean canInteractWith(PlayerEntity player) {
+	public boolean canInteractWith(Player player) {
 		if(level.getBlockEntity(worldPosition) != this)return false;
 		double dist = ItemWirelessTerminal.isPlayerHolding(player) ? Config.wirelessRange*2*Config.wirelessRange*2 : 64;
 		return !(player.distanceToSqr(this.worldPosition.getX() + 0.5D, this.worldPosition.getY() + 0.5D, this.worldPosition.getZ() + 0.5D) > dist);
@@ -143,15 +144,15 @@ public class TileEntityStorageTerminal extends TileEntity implements INamedConta
 	}
 
 	@Override
-	public CompoundNBT save(CompoundNBT compound) {
+	public CompoundTag save(CompoundTag compound) {
 		compound.putInt("sort", sort);
 		return super.save(compound);
 	}
 
 	@Override
-	public void load(BlockState st, CompoundNBT compound) {
+	public void load(CompoundTag compound) {
 		sort = compound.getInt("sort");
-		super.load(st, compound);
+		super.load(compound);
 	}
 
 	public String getLastSearch() {
